move `makeDragEvent` to `event-helpers.test-utils`
